[Library -> ViewModels] Prepare GyroShell for MVVM
