+++ GyroShell.Library/ViewModels/ModulesSettingViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GyroShell.Library.ViewModels
+{
+    public sealed class ModulesSettingViewModel
+    {
+    }
+}
