+++ GyroShell/Controls/SettingsWindow.xaml
@@ -1,0 +1,54 @@
+<Window
+    x:Class="GyroShell.Controls.SettingsWindow"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:GyroShell.Controls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="40"/>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+        <Grid x:Name="AppTitleBar" Grid.Row="0" Background="Transparent" Opacity="1">
+            <StackPanel Orientation="Horizontal">
+                <TextBlock FontFamily="Segoe MDL2 Assets" Text="&#xE713;" Margin="16,15,0,0" FontSize="18" x:Name="TopIcon"/>
+                <TextBlock Text="GyroShell Settings" Margin="16,16,0,0" Style="{StaticResource CaptionTextBlockStyle}" FontSize="12"/>
+            </StackPanel>
+        </Grid>
+        <Grid x:Name="ContentGrid" Grid.Row="1">
+            <NavigationView x:Name="SettingNav" IsSettingsVisible="False" IsBackButtonVisible="Collapsed" SelectionChanged="SettingNav_SelectionChanged">
+                <NavigationView.MenuItems>
+                    <NavigationViewItem Content="Customization" Tag="Customization">
+                        <NavigationViewItem.Icon>
+                            <FontIcon FontFamily="Segoe MDL2 Assets" Glyph="&#xEC87;"/>
+                        </NavigationViewItem.Icon>
+                    </NavigationViewItem>
+                    <NavigationViewItem Content="Appearance" Tag="Appearance">
+                        <NavigationViewItem.Icon>
+                            <FontIcon FontFamily="Segoe MDL2 Assets" Glyph="&#xE790;"/>
+                        </NavigationViewItem.Icon>
+                    </NavigationViewItem>
+                    <NavigationViewItem Content="Behavior" Tag="Behavior">
+                        <NavigationViewItem.Icon>
+                            <FontIcon FontFamily="Segoe MDL2 Assets" Glyph="&#xE790;"/>
+                        </NavigationViewItem.Icon>
+                    </NavigationViewItem>
+                    <NavigationViewItem Content="Modules" Tag="Modules">
+                        <NavigationViewItem.Icon>
+                            <FontIcon FontFamily="Segoe MDL2 Assets" Glyph="&#xED2A;"/>
+                        </NavigationViewItem.Icon>
+                    </NavigationViewItem>
+                    <NavigationViewItem Content="About" Tag="AboutPage">
+                        <NavigationViewItem.Icon>
+                            <FontIcon FontFamily="Segoe MDL2 Assets" Glyph="&#xE946;"/>
+                        </NavigationViewItem.Icon>
+                    </NavigationViewItem>
+                </NavigationView.MenuItems>
+                <Frame x:Name="contentFrame"/>
+            </NavigationView>
+        </Grid>
+    </Grid>
+</Window>
