+++ GyroShell/GyroShell.csproj
@@ -15,13 +15,9 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="Commands\**" />
-    <Compile Remove="NewFolder\**" />
     <EmbeddedResource Remove="Commands\**" />
-    <EmbeddedResource Remove="NewFolder\**" />
     <None Remove="Commands\**" />
-    <None Remove="NewFolder\**" />
     <Page Remove="Commands\**" />
-    <Page Remove="NewFolder\**" />
   </ItemGroup>
 
   <ItemGroup>
@@ -34,6 +30,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="CommunityToolkit.Mvvm" Version="8.2.1" />
     <PackageReference Include="CommunityToolkit.WinUI.Connectivity" Version="7.1.2" />
     <PackageReference Include="CoreAudio" Version="1.37.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
@@ -72,17 +69,11 @@
     <None Remove="Controls\StartupScreen.xaml" />
     <None Remove="Controls\StartWindow.xaml" />
     <None Remove="Settings\AboutPage.xaml" />
-    <None Remove="Settings\Customization.xaml" />
-    <None Remove="Settings\MoudlesPage.xaml" />
     <None Remove="Settings\SettingsWindow.xaml" />
+    <None Remove="Views\ModulesSettingView.xaml" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\GyroShell.Library\GyroShell.Library.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <Page Update="Settings\MoudlesPage.xaml">
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
   </ItemGroup>
   <ItemGroup>
     <Page Update="Controls\StartupScreen.xaml">
@@ -115,6 +106,8 @@
     </Page>
   </ItemGroup>
   <ItemGroup>
-    <PRIResource Remove="NewFolder\**" />
+    <Page Update="Views\ModulesSettingView.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
 </Project>
