+++ GyroShell/Views/ModulesSettingView.xaml
@@ -1,8 +1,8 @@
 <Page
-    x:Class="GyroShell.Settings.MoudlesPage"
+    x:Class="GyroShell.Views.ModulesSettingView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:GyroShell.Settings"
+    xmlns:local="using:GyroShell.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
